Update initialization.py
specify int64 for long to avoid default int32  of win64 numpy
as mentioned in
https://stackoverflow.com/questions/36797420/easiest-way-to-generate-random-int64-array-in-numpy

+++ jiant/shared/initialization.py
@@ -187,7 +187,7 @@ def get_seed(seed):
 
     """
     if seed == -1:
-        return int(np.random.randint(0, 2 ** 32 - 1))
+        return int(np.random.randint(0, 2 ** 32 - 1,dtype=np.int64))
     else:
         return seed
